requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
croniter==6.0.0
crontab==1.0.5
dj-rest-auth==5.0.2
Django==6.0
django-allauth==65.13.1
django-cors-headers==4.9.0
django-rq==3.2.2
django-stubs-ext==5.2.8
django-tasks==0.10.0
djangorestframework==3.16.1
freezegun==1.5.5
h11==0.16.0
idna==3.11
iniconfig==2.3.0
mysql-connector-python==9.5.0
mysqlclient==2.2.7
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
rq==2.6.1
rq-scheduler==0.14.0
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0
